+++ apps/foh/src/app.routes.tsx
@@ -1,23 +1,32 @@
-import { RouteObject } from 'react-router-dom';
+import { MakeGenerics, Route } from '@tanstack/react-location';
 
-import Layout from '@/app/layout';
+import { Championship } from 'lib';
+
+import Layout, { rootLoader } from '@/app/layout';
 
 import HomePage from '@/app/home/page';
 import PlayersPage from '@/app/players/page';
 import RankingPage from '@/app/ranking/page';
 import MatchesPage from '@/app/matches/page';
+import { SplashScreen } from 'ui';
 
-const appRoutes: RouteObject[] = [
+export type LocationGenerics = MakeGenerics<{
+  LoaderData: {
+    championships: Championship[];
+    currentChampionship: Championship | null;
+  };
+}>;
+
+export const appRoutes: Route<LocationGenerics>[] = [
   {
-    path: '/',
     element: <Layout />,
+    loader: rootLoader,
+    pendingElement: async () => <SplashScreen />,
     children: [
-      { index: true, element: <HomePage /> },
+      { path: '/', element: <HomePage /> },
       { path: 'tabelle', element: <RankingPage /> },
       { path: 'spieler', element: <PlayersPage /> },
       { path: 'spiele', element: <MatchesPage /> },
     ],
   },
 ];
-
-export default appRoutes;